+++ src/main/java/com/example/dbms_group2/model/DTO/VendorDetailInsideDTO.java
@@ -3,13 +3,14 @@ package com.example.dbms_group2.model.DTO;
 public class VendorDetailInsideDTO {
     private String title;                 // 攤商標題
     private String description;// 攤商介紹
-    private int applyId;
+    private Long applyId;
 
-    public VendorDetailInsideDTO(String title, String description, int applyId) {
+    public VendorDetailInsideDTO(String title, String description, Long applyId) {
         this.title = title;
         this.description = description;
         this.applyId = applyId;
     }
+
 
     public String getTitle() {
         return title;
@@ -27,11 +28,11 @@ public class VendorDetailInsideDTO {
         this.description = description;
     }
 
-    public int getApplyId() {
+    public Long getApplyId() {
         return applyId;
     }
 
-    public void setApplyId(int applyId) {
+    public void setApplyId(Long applyId) {
         this.applyId = applyId;
     }
 }
